+++ src/realm/object-store/c_api/notifications.cpp
@@ -128,6 +128,22 @@ RLM_API realm_notification_token_t* realm_list_add_notification_callback(realm_l
     });
 }
 
+RLM_API realm_notification_token_t* realm_set_add_notification_callback(realm_set_t* set, void* userdata,
+                                                                        realm_free_userdata_func_t free,
+                                                                        realm_on_collection_change_func_t on_change,
+                                                                        realm_callback_error_func_t on_error,
+                                                                        realm_scheduler_t*)
+{
+    return wrap_err([&]() {
+        CollectionNotificationsCallback cb;
+        cb.m_userdata = UserdataPtr{userdata, free};
+        cb.m_on_change = on_change;
+        cb.m_on_error = on_error;
+        auto token = set->add_notification_callback(std::move(cb));
+        return new realm_notification_token_t{std::move(token)};
+    });
+}
+
 RLM_API realm_notification_token_t*
 realm_results_add_notification_callback(realm_results_t* results, void* userdata, realm_free_userdata_func_t free,
                                         realm_on_collection_change_func_t on_change,