+++ test/object-store/c_api/c_api.cpp
@@ -447,6 +447,15 @@ std::vector<realm_property_info_t> all_property_types(const char* link_target)
         "nullable_timestamp_list", "nullable_float_list", "nullable_double_list", "nullable_decimal_list",
         "nullable_object_id_list", "nullable_uuid_list",
     };
+    static const char* set_names[] = {
+        "int_set",   "bool_set",   "string_set",  "binary_set",    "timestamp_set",
+        "float_set", "double_set", "decimal_set", "object_id_set", "uuid_set",
+    };
+    static const char* nullable_set_names[] = {
+        "nullable_int_set",       "nullable_bool_set",  "nullable_string_set", "nullable_binary_set",
+        "nullable_timestamp_set", "nullable_float_set", "nullable_double_set", "nullable_decimal_set",
+        "nullable_object_id_set", "nullable_uuid_set",
+    };
     static const realm_property_type_e types[] = {
         RLM_PROPERTY_TYPE_INT,       RLM_PROPERTY_TYPE_BOOL,  RLM_PROPERTY_TYPE_STRING, RLM_PROPERTY_TYPE_BINARY,
         RLM_PROPERTY_TYPE_TIMESTAMP, RLM_PROPERTY_TYPE_FLOAT, RLM_PROPERTY_TYPE_DOUBLE, RLM_PROPERTY_TYPE_DECIMAL128,
@@ -457,12 +466,16 @@ std::vector<realm_property_info_t> all_property_types(const char* link_target)
     size_t num_nullable_names = std::distance(std::begin(nullable_names), std::end(nullable_names));
     size_t num_list_names = std::distance(std::begin(list_names), std::end(list_names));
     size_t num_nullable_list_names = std::distance(std::begin(nullable_list_names), std::end(nullable_list_names));
+    size_t num_set_names = std::distance(std::begin(set_names), std::end(set_names));
+    size_t num_nullable_set_names = std::distance(std::begin(nullable_set_names), std::end(nullable_set_names));
     size_t num_types = std::distance(std::begin(types), std::end(types));
 
     REALM_ASSERT(num_names == num_types);
     REALM_ASSERT(num_nullable_names == num_types);
     REALM_ASSERT(num_list_names == num_types);
     REALM_ASSERT(num_nullable_list_names == num_types);
+    REALM_ASSERT(num_set_names == num_types);
+    REALM_ASSERT(num_nullable_set_names == num_types);
 
     for (size_t i = 0; i < num_names; ++i) {
         const char* public_name = i == 0 ? "public_int" : "";
@@ -488,11 +501,21 @@ std::vector<realm_property_info_t> all_property_types(const char* link_target)
             nullable_list_names[i], "", types[i], RLM_COLLECTION_TYPE_LIST, "", "", RLM_INVALID_PROPERTY_KEY,
             RLM_PROPERTY_NULLABLE,
         };
+        realm_property_info_t set{
+            set_names[i],        "", types[i], RLM_COLLECTION_TYPE_SET, "", "", RLM_INVALID_PROPERTY_KEY,
+            RLM_PROPERTY_NORMAL,
+        };
+        realm_property_info_t nullable_set{
+            nullable_set_names[i], "", types[i], RLM_COLLECTION_TYPE_SET, "", "", RLM_INVALID_PROPERTY_KEY,
+            RLM_PROPERTY_NULLABLE,
+        };
 
         properties.push_back(normal);
         properties.push_back(nullable);
         properties.push_back(list);
         properties.push_back(nullable_list);
+        properties.push_back(set);
+        properties.push_back(nullable_set);
     }
 
     realm_property_info_t link{
@@ -503,9 +526,14 @@ std::vector<realm_property_info_t> all_property_types(const char* link_target)
         "link_list", "", RLM_PROPERTY_TYPE_OBJECT, RLM_COLLECTION_TYPE_LIST,
         link_target, "", RLM_INVALID_PROPERTY_KEY, RLM_PROPERTY_NORMAL,
     };
+    realm_property_info_t link_set{
+        "link_set",  "", RLM_PROPERTY_TYPE_OBJECT, RLM_COLLECTION_TYPE_SET,
+        link_target, "", RLM_INVALID_PROPERTY_KEY, RLM_PROPERTY_NORMAL,
+    };
 
     properties.push_back(link);
     properties.push_back(link_list);
+    properties.push_back(link_set);
 
     // realm_property_info_t mixed{
     //     "mixed", "", RLM_PROPERTY_TYPE_MIXED,  RLM_COLLECTION_TYPE_NONE,
@@ -2224,6 +2252,413 @@ TEST_CASE("C API") {
             }
         }
 
+        SECTION("sets") {
+            SECTION("realm_get_set() errors") {
+                CHECK(!realm_get_set(obj1.get(), foo_properties["int"]));
+                CHECK_ERR(RLM_ERR_PROPERTY_TYPE_MISMATCH);
+
+                CHECK(!realm_get_set(obj1.get(), 123123123));
+                CHECK_ERR(RLM_ERR_INVALID_PROPERTY);
+            }
+
+            SECTION("nullable strings") {
+                auto strings = cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_string_set"]));
+                CHECK(strings);
+                CHECK(!realm_is_frozen(strings.get()));
+
+                realm_value_t a = rlm_str_val("a");
+                realm_value_t b = rlm_str_val("b");
+                realm_value_t c = rlm_null();
+
+                SECTION("realm_equals() type check") {
+                    CHECK(!realm_equals(strings.get(), obj1.get()));
+                }
+
+                SECTION("realm_clone()") {
+                    auto set2 = clone_cptr(strings.get());
+                    CHECK(realm_equals(strings.get(), set2.get()));
+                    CHECK(strings.get() != set2.get());
+                }
+
+                SECTION("insert, then get, then erase") {
+                    write([&]() {
+                        bool inserted = false;
+                        CHECK(checked(realm_set_insert(strings.get(), a, nullptr, &inserted)));
+                        CHECK(inserted);
+                        CHECK(checked(realm_set_insert(strings.get(), b, nullptr, &inserted)));
+                        CHECK(inserted);
+                        CHECK(checked(realm_set_insert(strings.get(), c, nullptr, &inserted)));
+                        CHECK(inserted);
+
+                        size_t a_index, b_index, c_index;
+                        bool found = false;
+                        CHECK(checked(realm_set_find(strings.get(), a, &a_index, &found)));
+                        CHECK(found);
+                        CHECK(checked(realm_set_find(strings.get(), b, &b_index, &found)));
+                        CHECK(found);
+                        CHECK(checked(realm_set_find(strings.get(), c, &c_index, &found)));
+                        CHECK(found);
+
+                        realm_value_t a2, b2, c2;
+                        CHECK(checked(realm_set_get(strings.get(), a_index, &a2)));
+                        CHECK(checked(realm_set_get(strings.get(), b_index, &b2)));
+                        CHECK(checked(realm_set_get(strings.get(), c_index, &c2)));
+
+                        CHECK(rlm_stdstr(a2) == "a");
+                        CHECK(rlm_stdstr(b2) == "b");
+                        CHECK(c2.type == RLM_TYPE_NULL);
+
+                        bool erased = false;
+                        CHECK(checked(realm_set_erase(strings.get(), a2, &erased)));
+                        CHECK(erased);
+                        CHECK(checked(realm_set_erase(strings.get(), rlm_int_val(987), &erased)));
+                        CHECK(!erased);
+                    });
+                }
+
+                SECTION("equality") {
+                    auto strings2 = cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_string_set"]));
+                    CHECK(strings2);
+                    CHECK(realm_equals(strings.get(), strings2.get()));
+
+                    write([&]() {
+                        auto obj3 = cptr_checked(realm_object_create(realm, class_foo.key));
+                        CHECK(obj3);
+                        auto strings3 =
+                            cptr_checked(realm_get_set(obj3.get(), foo_properties["nullable_string_set"]));
+                        CHECK(!realm_equals(strings.get(), strings3.get()));
+                    });
+                }
+            }
+
+            SECTION("get/insert all property types") {
+                realm_value_t null = rlm_null();
+                realm_value_t integer = rlm_int_val(987);
+                realm_value_t boolean = rlm_bool_val(true);
+                realm_value_t string = rlm_str_val("My string");
+                const uint8_t binary_data[] = {0, 1, 2, 3, 4, 5, 6, 7};
+                realm_value_t binary = rlm_binary_val(binary_data, sizeof(binary_data));
+                realm_value_t timestamp = rlm_timestamp_val(1000000, 123123123);
+                realm_value_t fnum = rlm_float_val(123.f);
+                realm_value_t dnum = rlm_double_val(456.0);
+                realm_value_t decimal = rlm_decimal_val(999.0);
+                realm_value_t object_id = rlm_object_id_val("abc123abc123");
+                realm_value_t uuid = rlm_uuid_val("01234567-9abc-4def-9012-3456789abcde");
+
+                auto int_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["int_set"]));
+                auto bool_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["bool_set"]));
+                auto string_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["string_set"]));
+                auto binary_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["binary_set"]));
+                auto timestamp_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["timestamp_set"]));
+                auto float_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["float_set"]));
+                auto double_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["double_set"]));
+                auto decimal_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["decimal_set"]));
+                auto object_id_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["object_id_set"]));
+                auto uuid_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["uuid_set"]));
+                auto nullable_int_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_int_set"]));
+                auto nullable_bool_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_bool_set"]));
+                auto nullable_string_set =
+                    cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_string_set"]));
+                auto nullable_binary_set =
+                    cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_binary_set"]));
+                auto nullable_timestamp_set =
+                    cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_timestamp_set"]));
+                auto nullable_float_set =
+                    cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_float_set"]));
+                auto nullable_double_set =
+                    cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_double_set"]));
+                auto nullable_decimal_set =
+                    cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_decimal_set"]));
+                auto nullable_object_id_set =
+                    cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_object_id_set"]));
+                auto nullable_uuid_set = cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_uuid_set"]));
+
+                write([&]() {
+                    CHECK(realm_set_insert(int_set.get(), integer, nullptr, nullptr));
+                    CHECK(realm_set_insert(bool_set.get(), boolean, nullptr, nullptr));
+                    CHECK(realm_set_insert(string_set.get(), string, nullptr, nullptr));
+                    CHECK(realm_set_insert(binary_set.get(), binary, nullptr, nullptr));
+                    CHECK(realm_set_insert(timestamp_set.get(), timestamp, nullptr, nullptr));
+                    CHECK(realm_set_insert(float_set.get(), fnum, nullptr, nullptr));
+                    CHECK(realm_set_insert(double_set.get(), dnum, nullptr, nullptr));
+                    CHECK(realm_set_insert(decimal_set.get(), decimal, nullptr, nullptr));
+                    CHECK(realm_set_insert(object_id_set.get(), object_id, nullptr, nullptr));
+                    CHECK(realm_set_insert(uuid_set.get(), uuid, nullptr, nullptr));
+
+                    CHECK(realm_set_insert(nullable_int_set.get(), integer, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_bool_set.get(), boolean, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_string_set.get(), string, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_binary_set.get(), binary, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_timestamp_set.get(), timestamp, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_float_set.get(), fnum, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_double_set.get(), dnum, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_decimal_set.get(), decimal, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_object_id_set.get(), object_id, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_uuid_set.get(), uuid, nullptr, nullptr));
+
+                    CHECK(realm_set_insert(nullable_int_set.get(), null, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_bool_set.get(), null, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_string_set.get(), null, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_binary_set.get(), null, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_timestamp_set.get(), null, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_float_set.get(), null, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_double_set.get(), null, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_decimal_set.get(), null, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_object_id_set.get(), null, nullptr, nullptr));
+                    CHECK(realm_set_insert(nullable_uuid_set.get(), null, nullptr, nullptr));
+                });
+
+                realm_value_t value;
+
+                CHECK(realm_set_get(int_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, integer));
+                CHECK(realm_set_get(bool_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, boolean));
+                CHECK(realm_set_get(string_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, string));
+                CHECK(realm_set_get(binary_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, binary));
+                CHECK(realm_set_get(timestamp_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, timestamp));
+                CHECK(realm_set_get(float_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, fnum));
+                CHECK(realm_set_get(double_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, dnum));
+                CHECK(realm_set_get(decimal_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, decimal));
+                CHECK(realm_set_get(object_id_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, object_id));
+                CHECK(realm_set_get(uuid_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, uuid));
+                CHECK(realm_set_get(nullable_int_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, integer));
+                CHECK(realm_set_get(nullable_bool_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, boolean));
+                CHECK(realm_set_get(nullable_string_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, string));
+                CHECK(realm_set_get(nullable_binary_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, binary));
+                CHECK(realm_set_get(nullable_timestamp_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, timestamp));
+                CHECK(realm_set_get(nullable_float_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, fnum));
+                CHECK(realm_set_get(nullable_double_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, dnum));
+                CHECK(realm_set_get(nullable_decimal_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, decimal));
+                CHECK(realm_set_get(nullable_object_id_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, object_id));
+                CHECK(realm_set_get(nullable_uuid_set.get(), 1, &value));
+                CHECK(rlm_val_eq(value, uuid));
+
+                write([&]() {
+                    size_t index;
+                    bool inserted;
+                    CHECK(realm_set_insert(nullable_int_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                    CHECK(realm_set_insert(nullable_bool_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                    CHECK(realm_set_insert(nullable_string_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                    CHECK(realm_set_insert(nullable_binary_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                    CHECK(realm_set_insert(nullable_timestamp_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                    CHECK(realm_set_insert(nullable_float_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                    CHECK(realm_set_insert(nullable_double_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                    CHECK(realm_set_insert(nullable_decimal_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                    CHECK(realm_set_insert(nullable_object_id_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                    CHECK(realm_set_insert(nullable_uuid_set.get(), null, &index, &inserted));
+                    CHECK((index == 0 && !inserted));
+                });
+
+                // Note: This relies on the fact that NULL is "less than" other
+                // values in the internal sort order.
+                CHECK(realm_set_get(nullable_int_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+                CHECK(realm_set_get(nullable_bool_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+                CHECK(realm_set_get(nullable_string_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+                CHECK(realm_set_get(nullable_binary_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+                CHECK(realm_set_get(nullable_timestamp_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+                CHECK(realm_set_get(nullable_float_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+                CHECK(realm_set_get(nullable_double_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+                CHECK(realm_set_get(nullable_decimal_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+                CHECK(realm_set_get(nullable_object_id_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+                CHECK(realm_set_get(nullable_uuid_set.get(), 0, &value));
+                CHECK(rlm_val_eq(value, null));
+            }
+
+            SECTION("links") {
+                CPtr<realm_set_t> bars;
+
+                write([&]() {
+                    bars = cptr_checked(realm_get_set(obj1.get(), foo_properties["link_set"]));
+                    auto bar_link = realm_object_as_link(obj2.get());
+                    realm_value_t bar_link_val;
+                    bar_link_val.type = RLM_TYPE_LINK;
+                    bar_link_val.link = bar_link;
+                    size_t index;
+                    bool inserted;
+                    CHECK(checked(realm_set_insert(bars.get(), bar_link_val, &index, &inserted)));
+                    CHECK((index == 0 && inserted));
+                    CHECK(checked(realm_set_insert(bars.get(), bar_link_val, &index, &inserted)));
+                    CHECK((index == 0 && !inserted));
+                    size_t size;
+                    CHECK(checked(realm_set_size(bars.get(), &size)));
+                    CHECK(size == 1);
+                });
+
+                SECTION("get") {
+                    realm_value_t val;
+                    CHECK(checked(realm_set_get(bars.get(), 0, &val)));
+                    CHECK(val.type == RLM_TYPE_LINK);
+                    CHECK(val.link.target_table == class_bar.key);
+                    CHECK(val.link.target == realm_object_get_key(obj2.get()));
+
+                    auto result = realm_set_get(bars.get(), 1, &val);
+                    CHECK(!result);
+                    CHECK_ERR(RLM_ERR_INDEX_OUT_OF_BOUNDS);
+                }
+
+                SECTION("insert wrong type") {
+                    write([&]() {
+                        auto foo2 = cptr(realm_object_create(realm, class_foo.key));
+                        CHECK(foo2);
+                        realm_value_t foo2_link_val;
+                        foo2_link_val.type = RLM_TYPE_LINK;
+                        foo2_link_val.link = realm_object_as_link(foo2.get());
+
+                        CHECK(!realm_set_insert(bars.get(), foo2_link_val, nullptr, nullptr));
+                        CHECK_ERR(RLM_ERR_PROPERTY_TYPE_MISMATCH);
+                    });
+                }
+
+                SECTION("realm_set_clear()") {
+                    write([&]() {
+                        CHECK(realm_set_clear(bars.get()));
+                    });
+                    size_t size;
+                    CHECK(realm_set_size(bars.get(), &size));
+                    CHECK(size == 0);
+
+                    size_t num_bars;
+                    CHECK(realm_get_num_objects(realm, class_bar.key, &num_bars));
+                    CHECK(num_bars != 0);
+                }
+
+                SECTION("realm_set_remove_all()") {
+                    realm_value_t val;
+                    CHECK(checked(realm_set_get(bars.get(), 0, &val)));
+                    CHECK(val.type == RLM_TYPE_LINK);
+                    CHECK(val.link.target_table == class_bar.key);
+                    CHECK(val.link.target == realm_object_get_key(obj2.get()));
+
+                    size_t num_bars;
+                    size_t size;
+
+                    write([&]() {
+                        CHECK(checked(realm_set_remove_all(bars.get())));
+                    });
+
+                    CHECK(realm_set_size(bars.get(), &size));
+                    CHECK(size == 0);
+
+                    CHECK(realm_get_num_objects(realm, class_bar.key, &num_bars));
+                    CHECK(num_bars == 0);
+                }
+            }
+
+            SECTION("notifications") {
+                struct State {
+                    CPtr<realm_collection_changes_t> changes;
+                    CPtr<realm_async_error_t> error;
+                    bool destroyed = false;
+                };
+
+                State state;
+
+                auto on_change = [](void* userdata, const realm_collection_changes_t* changes) {
+                    auto* state = static_cast<State*>(userdata);
+                    state->changes = clone_cptr(changes);
+                };
+
+                auto on_error = [](void* userdata, const realm_async_error_t* err) {
+                    auto* state = static_cast<State*>(userdata);
+                    state->error = clone_cptr(err);
+                };
+
+                CPtr<realm_set_t> strings =
+                    cptr_checked(realm_get_set(obj1.get(), foo_properties["nullable_string_set"]));
+
+                auto str1 = rlm_str_val("a");
+                auto str2 = rlm_str_val("b");
+                auto null = rlm_null();
+
+                auto require_change = [&]() {
+                    auto token = cptr_checked(realm_set_add_notification_callback(strings.get(), &state, nullptr,
+                                                                                  on_change, on_error, nullptr));
+                    checked(realm_refresh(realm));
+                    return token;
+                };
+
+                SECTION("userdata is freed when the token is destroyed") {
+                    auto token = cptr_checked(realm_set_add_notification_callback(
+                        strings.get(), &state,
+                        [](void* p) {
+                            static_cast<State*>(p)->destroyed = true;
+                        },
+                        nullptr, nullptr, nullptr));
+                    CHECK(!state.destroyed);
+                    token.reset();
+                    CHECK(state.destroyed);
+                }
+
+                SECTION("insertion,deletion sends a change callback") {
+                    write([&]() {
+                        checked(realm_set_insert(strings.get(), str1, nullptr, nullptr));
+                    });
+
+                    auto token = require_change();
+                    write([&]() {
+                        checked(realm_set_erase(strings.get(), str1, nullptr));
+                        checked(realm_set_insert(strings.get(), str2, nullptr, nullptr));
+                        checked(realm_set_insert(strings.get(), null, nullptr, nullptr));
+                    });
+                    CHECK(!state.error);
+                    CHECK(state.changes);
+
+                    size_t num_deletion_ranges, num_insertion_ranges, num_modification_ranges, num_moves;
+                    realm_collection_changes_get_num_ranges(state.changes.get(), &num_deletion_ranges,
+                                                            &num_insertion_ranges, &num_modification_ranges,
+                                                            &num_moves);
+                    CHECK(num_deletion_ranges == 1);
+                    CHECK(num_insertion_ranges == 1);
+                    CHECK(num_modification_ranges == 0);
+                    CHECK(num_moves == 0);
+
+                    realm_index_range_t insertion_range, deletion_range;
+                    realm_collection_changes_get_ranges(state.changes.get(), &deletion_range, 1, &insertion_range, 1,
+                                                        nullptr, 0, nullptr, 0, nullptr, 0);
+                    CHECK(deletion_range.from == 0);
+                    CHECK(deletion_range.to == 1);
+                    CHECK(insertion_range.from == 0);
+                    CHECK(insertion_range.to == 2);
+                }
+            }
+        }
+
         SECTION("notifications") {
             struct State {
                 CPtr<realm_object_changes_t> changes;
@@ -2296,6 +2731,7 @@ TEST_CASE("C API") {
         });
 
         auto list = cptr_checked(realm_get_list(foo_obj.get(), foo_properties["int_list"]));
+        auto set = cptr_checked(realm_get_set(foo_obj.get(), foo_properties["int_set"]));
         auto results = cptr_checked(realm_object_find_all(realm, class_foo.key));
 
         SECTION("wrong thread") {
@@ -2312,6 +2748,7 @@ TEST_CASE("C API") {
             auto foo_obj_tsr = cptr_checked(realm_create_thread_safe_reference(foo_obj.get()));
             auto bar_obj_tsr = cptr_checked(realm_create_thread_safe_reference(bar_obj.get()));
             auto list_tsr = cptr_checked(realm_create_thread_safe_reference(list.get()));
+            auto set_tsr = cptr_checked(realm_create_thread_safe_reference(set.get()));
             auto results_tsr = cptr_checked(realm_create_thread_safe_reference(results.get()));
 
             SECTION("resolve") {
@@ -2324,6 +2761,8 @@ TEST_CASE("C API") {
                         cptr_checked(realm_object_from_thread_safe_reference(realm2.get(), bar_obj_tsr.get()));
                     auto results2 =
                         cptr_checked(realm_results_from_thread_safe_reference(realm2.get(), results_tsr.get()));
+                    auto list2 = cptr_checked(realm_list_from_thread_safe_reference(realm2.get(), list_tsr.get()));
+                    auto set2 = cptr_checked(realm_set_from_thread_safe_reference(realm2.get(), set_tsr.get()));
 
                     realm_value_t foo_obj_int;
                     CHECK(realm_get_value(foo_obj2.get(), foo_int_key, &foo_obj_int));
@@ -2351,6 +2790,8 @@ TEST_CASE("C API") {
                 CHECK(!realm_object_from_thread_safe_reference(realm, list_tsr.get()));
                 CHECK_ERR(RLM_ERR_LOGIC);
                 CHECK(!realm_list_from_thread_safe_reference(realm, foo_obj_tsr.get()));
+                CHECK_ERR(RLM_ERR_LOGIC);
+                CHECK(!realm_set_from_thread_safe_reference(realm, list_tsr.get()));
                 CHECK_ERR(RLM_ERR_LOGIC);
                 CHECK(!realm_results_from_thread_safe_reference(realm, list_tsr.get()));
                 CHECK_ERR(RLM_ERR_LOGIC);
@@ -2483,13 +2924,59 @@ TEST_CASE("C API") {
             realm_list_size(thawed_list, &count);
             CHECK(count == 1);
 
+            CHECK(realm_list_is_valid(thawed_list));
             write([&]() {
                 CHECK(checked(realm_object_delete(obj1.get())));
             });
+            CHECK(!realm_list_is_valid(thawed_list));
             realm_release(thawed_list);
             CHECK(realm_list_resolve_in(frozen_list, realm, &thawed_list));
             CHECK(thawed_list == nullptr);
             realm_release(frozen_list);
+        }
+
+        SECTION("sets") {
+            CPtr<realm_object_t> obj1;
+            size_t count;
+
+            write([&]() {
+                obj1 = cptr_checked(realm_object_create(realm, class_foo.key));
+                CHECK(obj1);
+            });
+
+            auto set = cptr_checked(realm_get_set(obj1.get(), foo_properties["string_set"]));
+            realm_set_size(set.get(), &count);
+            CHECK(count == 0);
+
+            auto frozen_realm = cptr_checked(realm_freeze(realm));
+            realm_set_t* frozen_set;
+            CHECK(realm_set_resolve_in(set.get(), frozen_realm.get(), &frozen_set));
+            realm_set_size(frozen_set, &count);
+            CHECK(count == 0);
+
+            write([&]() {
+                checked(realm_set_insert(set.get(), rlm_str_val("Hello"), nullptr, nullptr));
+            });
+
+            realm_set_size(frozen_set, &count);
+            CHECK(count == 0);
+            realm_set_size(set.get(), &count);
+            CHECK(count == 1);
+
+            realm_set_t* thawed_set;
+            CHECK(realm_set_resolve_in(frozen_set, realm, &thawed_set));
+            realm_set_size(thawed_set, &count);
+            CHECK(count == 1);
+
+            CHECK(realm_set_is_valid(thawed_set));
+            write([&]() {
+                CHECK(checked(realm_object_delete(obj1.get())));
+            });
+            CHECK(!realm_set_is_valid(thawed_set));
+            realm_release(thawed_set);
+            CHECK(realm_set_resolve_in(frozen_set, realm, &thawed_set));
+            CHECK(thawed_set == nullptr);
+            realm_release(frozen_set);
         }
     }