+++ test/test_sync_subscriptions.cpp
@@ -358,14 +358,14 @@ TEST(Sync_SubscriptionStoreNotifications)
     sub_set.update_state(SubscriptionSet::State::Complete);
     std::move(sub_set).commit();
 
-    CHECK_EQUAL(notification_futures[4].get(), SubscriptionSet::State::Superceded);
+    CHECK_EQUAL(notification_futures[4].get(), SubscriptionSet::State::Superseded);
     CHECK_EQUAL(notification_futures[5].get(), SubscriptionSet::State::Complete);
 
     // Also check that new requests for the superceded sub set get filled immediately.
     CHECK_EQUAL(old_sub_set.get_state_change_notification(SubscriptionSet::State::Complete).get(),
-                SubscriptionSet::State::Superceded);
+                SubscriptionSet::State::Superseded);
     old_sub_set.refresh();
-    CHECK_EQUAL(old_sub_set.state(), SubscriptionSet::State::Superceded);
+    CHECK_EQUAL(old_sub_set.state(), SubscriptionSet::State::Superseded);
 
     // Check that asking for a state change that is less than the current state of the sub set gets filled
     // immediately.